A message passing approach for decision fusion of hidden-markov observations in the presence of synchronized attacks
This code is is used to generate the results presented in the paper "A message passing approach for decision fusion of hidden-markov observations in the presence of synchronized attacks" presented at the Int. Conf. Advances in Multimedia (MMEDIA), Special Track on Models and Algorithms for Spatially and Temporally Correlated Data (STCD). The program has two types of file code for synchronized and desynchronized attacks that can be differentiated when the file end with "sync". To run a simulation it is enough to go the Main_sim_turbo_synch_biz_improved.m file and set the parameters. The result of the files are to calculate the messages on the message passing graph and to generate the markovian observation available to sensors in the network.